+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Changelog
 * Add `--email` option for `user:update` command to set email address of a user.
 * Add `--ask-password` option for `user:update` command to ask for password
 * Option `sugarcrm.path` is now relative to the configuration file instead of current directory.
+* Search for `.sugarclirc` files in all parent folders of the current directory. Configuration will
+  be overriden by the children.
 
 1.9.1
 ----